document cell component and provide defaults
